Data.Matchable

Matchable class

class (Eq1 t, Functor t) => Matchable t where

Containers that allows exact structural matching of two containers.

Minimal complete definition

zipMatchWith

Methods

zipMatch :: t a -> t b -> Maybe (t (a, b))

Decides if two structures match exactly. If they match, return zipped version of them.

zipMatch ta tb = Just tab

holds if and only if both of

ta = fmap fst tab
tb = fmap snd tab

holds. Otherwise, zipMatch ta tb = Nothing.

For example, the type signature of zipMatch on the list Functor [] reads as follows:

zipMatch :: [a] -> [b] -> Maybe [(a,b)]

zipMatch as bs returns Just (zip as bs) if the lengths of two given lists are same, and returns Nothing otherwise.

Example

>>> zipMatch [1, 2, 3] ['a', 'b', 'c']
Just [(1,'a'),(2,'b'),(3,'c')]
>>> zipMatch [1, 2, 3] ['a', 'b']
Nothing

zipMatchWith :: (a -> b -> Maybe c) -> t a -> t b -> Maybe (t c)

Match two structures. If they match, zip them with given function (a -> b -> Maybe c). Passed function can make whole match fail by returning Nothing.

A definition of zipMatchWith must satisfy:

If there is a pair (tab, tc) such that fulfills all following three conditions, then zipMatchWith f ta tb = Just tc.

ta = fmap fst tab

tb = fmap snd tab

fmap (uncurry f) tab = fmap Just tc

If there are no such pair, zipMatchWith f ta tb = Nothing.

If t is also Traversable, the last condition can be dropped and the equation can be stated without using tc.

zipMatchWith f ta tb = traverse (uncurry f) tab

zipMatch can be defined in terms of zipMatchWith. And if t is also Traversable, zipMatchWith can be defined in terms of zipMatch. When you implement both of them by hand, keep their relation in the way the default implementation is.

zipMatch             = zipMatchWith (curry pure)
zipMatchWith f ta tb = zipMatch ta tb >>= traverse (uncurry f)

zipzipMatch :: (Matchable t, Matchable u) => t (u a) -> t (u b) -> Maybe (t (u (a, b)))

zipzipMatch = zipMatchWith zipMatch

fmapRecovered :: Matchable t => (a -> b) -> t a -> t b

Matchable t implies Functor t. It is not recommended to implement fmap through this function, so it is named fmapRecovered but not fmapDefault.

eqDefault :: (Matchable t, Eq a) => t a -> t a -> Bool

Matchable t implies Eq a => Eq (t a).

liftEqDefault :: Matchable t => (a -> b -> Bool) -> t a -> t b -> Bool

Matchable t implies Eq1 t.

Define Matchable by Generic

class Matchable' t

An instance of Matchable can be implemened through GHC Generics. You only need to do two things: Make your type Functor and Generic1.

Example

>>> :set -XDeriveFunctor
>>> :set -XDeriveGeneric
>>> :{
  data MyTree label a = Leaf a | Node label [MyTree label a]
    deriving (Show, Read, Eq, Ord, Functor, Generic1)
:}

Then you can use genericZipMatchWith to implement zipMatchWith method. You also need Eq1 instance, but liftEqDefault is provided.

>>> :{
  instance (Eq label) => Matchable (MyTree label) where
    zipMatchWith = genericZipMatchWith
  instance (Eq label) => Eq1 (MyTree label) where
    liftEq = liftEqDefault
  :}

>>> zipMatch (Node "foo" [Leaf 1, Leaf 2]) (Node "foo" [Leaf 'a', Leaf 'b'])
Just (Node "foo" [Leaf (1,'a'),Leaf (2,'b')])
>>> zipMatch (Node "foo" [Leaf 1, Leaf 2]) (Node "bar" [Leaf 'a', Leaf 'b'])
Nothing
>>> zipMatch (Node "foo" [Leaf 1]) (Node "foo" [])
Nothing

Minimal complete definition

zipMatchWith'

genericZipMatchWith :: (Generic1 t, Matchable' (Rep1 t)) => (a -> b -> Maybe c) -> t a -> t b -> Maybe (t c)

zipMatchWith via Generics.